Lock parallel package operations (#2593)

+++ src/tox/util/path.py
@@ -4,10 +4,12 @@ from pathlib import Path
 from shutil import rmtree
 
 
-def ensure_empty_dir(path: Path) -> None:
+def ensure_empty_dir(path: Path, except_filename: str | None = None) -> None:
     if path.exists():
         if path.is_dir():
             for sub_path in path.iterdir():
+                if sub_path.name == except_filename:
+                    continue
                 if sub_path.is_dir():
                     rmtree(sub_path, ignore_errors=True)
                 else:
